+++ src/copilot_session_tools/scanner/__init__.py
@@ -5,7 +5,7 @@ Data structures are informed by:
 - microsoft/vscode (https://github.com/microsoft/vscode)
 """
 
-PARSER_VERSION = 4
+PARSER_VERSION = 7
 
 from .cli import _parse_cli_jsonl_file, _parse_workspace_yaml
 from .content import (